BUG/MINOR: quic: reject HANDSHAKE_DONE as server
As specified in RFC 9000, a client must never emit a HANDSHAKE_DONE
frame. If this happens, the server must close the connection with error
PROTOCOL VIOLATION.

Previously, such a frame was silently discarded on server side. The
connection remained opened which is not conformant to the specification.

This should be backported up to 2.6.

(cherry picked from commit cc29ab437ea6348aa1defcab15a306742a9fd826)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>