BUG/MINOR: quic: reject unknown frame type
As specified by RFC 9000, connection is closed on error if an unknown
QUIC frame type is received.

Previously, a frame with unknown type was silently discarded. The
connection remained opened which is not conformant to the specification.

This should be backported up to 2.6.

(cherry picked from commit 5a2aa8c1613eb83f3a7be7831a51869fe488e943)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>