BUG/MINOR: quic: fix output of show quic
Output of 'show quic' is messed up since the introduction of reordered
packets counter in the following commit. The new counter is mixed up
with the first stream line. This is due to the wrong placement of the
newline delimiter.

  167e38e0e0296e899aa894d2f3db5ba2a0c68cb5
  MINOR: quic: Add a counter for reordered packets

This should be backported up to 2.6.

(cherry picked from commit ba9f905da978afafb81ceb4277af2a886e6ca616)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>