BUG/MINOR: sink: fix a race condition in the TCP log forwarding code
That's exactly the same as commit 53bfab080c ("BUG/MINOR: sink: fix a race
condition between the writer and the reader") that went into 2.7 and was
backported as far as 2.4, except that since the code was duplicated, the
second instance was not noticed, leaving the race present. The race has
a limited impact, if a forwarder reaches the end of the logs and a new
message arrives before it leaves, the forwarder will only wake up after
yet another new message will be sent. In practice it remains unnoticeable
because for the race to trigger, one needs to have a steady flow of logs,
which means the wakeup will happen anyway.

This should be backported, but no need to insist on it if it resists.

(cherry picked from commit 962c129dc1f1f8718098e29ff9b01905fef63944)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 235961f8bcc7ababb5cb7e1e47452ccb606a76fc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>