MINOR: mux-h2: always use h2c_report_glitch()
The function aims at centralizing counter measures but due to the fact
that it only increments the counter by one unit, sometimes it was not
used and the value was calculated directly. Let's pass the increment in
argument so that it can be used everywhere.

(cherry picked from commit e6e7e1587e1d6ddc880b8ccdff09db41588e671e)
Signed-off-by: Willy Tarreau <w@1wt.eu>